requirements.txt
alembic==1.16.1
attrs==25.3.0
bleach==6.2.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
feedparser==6.0.11
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
markdown==3.7
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.9
Pygments==2.19.1
PySocks==1.7.1
pytest==8.4.0
pytest-flask==1.3.0
python-dotenv==1.1.0
python-http-client==3.3.7
requests==2.32.4
selenium==4.33.0
sendgrid==6.12.4
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
undetected-chromedriver==3.5.5
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
zipp==3.21.0
